split JS to dedicated pages + new buttons for edit and new trick
