+++ assets/css/app.scss
@@ -416,6 +416,15 @@ fieldset[disabled] .btn-primary.active {
     margin-left: 10px;
 }
 
+.edit-pencil {
+    margin-top : 130px;
+
+}
+
+.edit-pencil > a{
+    color : #f3f3f3;
+}
+
 /*-----------------------------------------------------------------------------------*/
