+++ src/Form/TrickFormType.php
@@ -29,21 +29,21 @@ class TrickFormType extends AbstractType
             ->add('category', EntityType::class, [
                 'class' => Category::class,
                 'choice_label' => 'name',
-                'placeholder' => 'Choisissez une catégorie'
+                'placeholder' => 'Choisissez une catégorie',
             ])
             ->add('images', CollectionType::class, [
                 'entry_type' => ImageFormType::class,
                 'entry_options' => ['label' => false],
                 'allow_add' => true,
                 'allow_delete' => true,
-                'by_reference' => false
+                'by_reference' => false,
             ])
             ->add('videos', CollectionType::class, [
                 'entry_type' => VideoFormType::class,
                 'entry_options' => ['label' => false],
                 'allow_add' => true,
                 'allow_delete' => true,
-                'by_reference' => false
+                'by_reference' => false,
             ])
             ;
     }
