+++ templates/trick.html.twig
@@ -10,8 +10,19 @@
         {% endif %}
             <div class="container">
                 <div class="row">       
-                    <div class="intro-text vertical-center smoothie col overlay">
-                        <div class="intro-heading wow fadeIn heading-font" data-wow-delay="0.2s"><h2>{{ trick.title }}</h2></div>                                   
+                    <div class="col overlay">
+                        <div class="row">
+                            <div class="col edit-pencil">
+                                <a href="{{ path('app_edit_trick', {id: trick.id}) }}" class="pull-right"><span class="fa fa-pencil fa-3x" aria-hidden="true"></span></a>
+                            </div>
+                        </div>
+                        <div class="row"> 
+                            <div class="col intro-text vertical-center smoothie">
+                                <div class="intro-heading wow fadeIn heading-font" data-wow-delay="0.2s">
+                                    <h2>{{ trick.title }}</h2>
+                                </div>                                 
+                            </div>                           
+                        </div>
                     </div>
                 </div>
             </div> 
@@ -22,6 +33,18 @@
             <div class="container">
                 <div class="row">
                     <div class="col col-carousel">
+                        {% for message in app.flashes('success') %}
+                            <div class="alert alert-success alert-dismissible" role="alert">
+                                <a href="#" class="close" data-dismiss="alert" aria-label="close">&times;</a>
+                                {{ message }}
+                            </div>    
+                        {% endfor %}
+                        {% for message in app.flashes('error') %}
+                            <div class="alert alert-danger alert-dismissible" role="alert">
+                                <a href="#" class="close" data-dismiss="alert" aria-label="close">&times;</a>
+                                {{ message }}
+                            </div>    
+                        {% endfor %}
             
                         {% if images is not empty %}
                             {% if images|length + videos|length == 1 %}
